rename alpaka examples
