+++ examples/alpaka/asynccopy/stb_image.h
