+++ examples/alpaka/asynccopy/stb_image_write.h
