+++ examples/alpaka/vectoradd/CMakeLists.txt
@@ -1,11 +1,12 @@
 cmake_minimum_required (VERSION 3.3)
-project(llama-asynccopy)
+project(llama-alpaka-vectoradd)
 
 if (NOT TARGET llama::llama)
 	find_package(llama REQUIRED)
 endif()
 find_package(alpaka 0.5.0 REQUIRED)
-ALPAKA_ADD_EXECUTABLE(${PROJECT_NAME} asynccopy.cpp ../common/Dummy.cpp stb_image.h stb_image_write.h)
+ALPAKA_ADD_EXECUTABLE(${PROJECT_NAME} vectoradd.cpp ../../common/Dummy.cpp)
 target_link_libraries(${PROJECT_NAME} PRIVATE llama::llama alpaka::alpaka)
 
-install(FILES "${PROJECT_BINARY_DIR}/llama-asynccopy" DESTINATION "bin" )
+install(FILES "${PROJECT_BINARY_DIR}/llama-alpaka-vectoradd" DESTINATION "bin")
+
